requirements.txt
gitpython>=3.1.41
binaryornot>=0.4.4
